Issue #000 fix: test cases fix.
